requirements.txt
appdirs==1.4.0
argh==0.26.2
backports-abc==0.5
brotlipy==0.6.0
certifi==2017.1.23
cffi==1.9.1
chardet==2.3.0
click==6.7
docutils==0.13.1
enum34==1.1.6
Flask==0.12
hapy-heritrix==0.1.1
idna==2.2
itsdangerous==0.24
Jinja2==2.9.5
lockfile==0.12.2
luigi==2.5.0
MarkupSafe==0.23
nose==1.3.7
packaging==16.8
pathtools==0.1.2
pycparser==2.17
pyparsing==2.1.10
python-daemon==2.1.2
pywb==0.33.0
PyYAML==3.12
redis==2.10.5
requests==2.13.0
requests-file==1.4.1
singledispatch==3.4.0.3
six==1.10.0
surt==0.3.0
tldextract==2.0.2
tornado==4.4.2
watchdog==0.8.3
webencodings==0.5
Werkzeug==0.11.15
